refactor utils HTTP infrastructure into package
