+++ utils/http/__init__.py
@@ -1,0 +1,20 @@
+# Copyright 2025-2026 Project N.E.K.O. Team
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+"""Shared HTTP clients, proxy policy, URL helpers, and TLS diagnostics.
+
+Implementations live in this package. The historical top-level ``utils``
+modules remain compatibility aliases so existing imports, singleton state,
+and module-private monkeypatch seams continue to refer to the same objects.
+"""
